animation fullPlayer

+++ components/Next-Ui/DeleteUserModal.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React, { useEffect } from "react";
 import {Modal, ModalContent, ModalHeader, ModalBody, ModalFooter, Button, useDisclosure, user} from "@nextui-org/react";
 import { BiUserCircle } from "react-icons/bi";
@@ -7,53 +8,47 @@ import useAuthModal from "@/hooks/useAuthModal";
 import { useUser } from "@/hooks/useUser";
 import toast from "react-hot-toast";
 
-export default function DeleteUserModal() {
-  const {isOpen, onOpen, onOpenChange} = useDisclosure();
-  
-  const router = useRouter()
-  const { supabaseClient } = useSessionContext()
-  const authModal = useAuthModal()
-  const { user } = useUser()
+export default function DeleteUserModal({ userData }: { userData: any }) {
+  const { isOpen, onOpen, onOpenChange } = useDisclosure();
+  const router = useRouter();
+  const { supabaseClient } = useSessionContext();
+  const authModal = useAuthModal();
+  const { user: currentUser } = useUser();
 
-
-// in development 
-
-    const handleDelete = async () => {
-        if (!user) {
-            return authModal.onOpen()
-          }
-   
-       
-      const { data, error } = await supabaseClient
-        .from("users")
-        .delete()
-        .eq("id", user.id )
-      
-
-        
-
-      if (!error && data) {
-       toast.success("yeah")
-      
+  const handleDelete = async () => {
+    try {
+      if (!currentUser) {
+        toast.error("Du bist nicht angemeldet.");
+        return;
       }
-    else if (error) {
-        console.log(error)
-    }}
 
+      const { error } = await supabaseClient
+      .from("users")
+      .delete()
+      .eq("id", currentUser.id);
+
+      if (!error) {
+        toast.success("Dein Konto wurde erfolgreich gelöscht.");
+        router.push("/dashboard");
+      } else {
+        throw new Error("Ein Fehler ist beim Löschen des Kontos aufgetreten.");
+      }
+    } catch (error) {
+      toast.error("Ein Fehler ist aufgetreten.");
+    }
+  };
 
   return (
     <>
-      <Button onPress={onOpen} className='w-80 lg:w-[30rem] ' color="danger" variant="bordered" startContent={<BiUserCircle size={26} />}>
+      <Button onPress={onOpen} className="w-80 lg:w-[30rem]" color="danger" variant="bordered" startContent={<BiUserCircle size={26} />}>
         Account löschen
-       </Button>
-      
+      </Button>
 
-      <Modal 
-  
-      className="absolute mx-auto "
-       placement="center"
-        backdrop="opaque" 
-        isOpen={isOpen} 
+      <Modal
+        className="absolute mx-auto"
+        placement="center"
+        backdrop="opaque"
+        isOpen={isOpen}
         onOpenChange={onOpenChange}
         motionProps={{
           variants: {
@@ -73,27 +68,24 @@ export default function DeleteUserModal() {
                 ease: "easeIn",
               },
             },
-          }
+          },
         }}
       >
-        <ModalContent  >
+        <ModalContent>
           {(onClose) => (
             <>
-              <ModalHeader  className="flex flex-col gap-1 text-xl">Account löschen</ModalHeader>
-              <ModalBody >
-                <p className=""> 
-                 Diese Aktion kann nicht rückgängig gemacht werden.
+              <ModalHeader className="flex flex-col gap-1 text-xl">Account löschen</ModalHeader>
+              <ModalBody>
+                <p className="">
+                  Diese Aktion kann nicht rückgängig gemacht werden.
                 </p>
-                <p>
-                 Nach dem löschen werden alle Daten aus dein Volume Profil endgütlig entfernt!
-                </p>
-            
+                <p>Nach dem Löschen werden alle Daten aus deinem Volume-Profil endgültig entfernt!</p>
               </ModalBody>
               <ModalFooter>
                 <Button color="danger" variant="light" onPress={onClose}>
                   Abbrechen
                 </Button>
-                <Button color="primary" onClick={handleDelete} >
+                <Button color="primary" onClick={handleDelete}>
                   Fortfahren
                 </Button>
               </ModalFooter>
@@ -101,7 +93,6 @@ export default function DeleteUserModal() {
           )}
         </ModalContent>
       </Modal>
-
     </>
   );
 }
